Update differential equation to 1D SWE
Why:
- Previous equations only modeled the effect of advection
- The new model represents a simplified version of 1D shallow-water
equations flowing through a channel whose cross-section is a rectangle
with constant width, assuming there is no shear stress.
How:
- Add a centered differencing function.
- Add a new variable to represent the flow's zonal velocity.
- Update simulation length and time delta to atenuate numerical
instability.
- Change real data types' precision to atenuate numerical instability.

+++ src/initial/mod_initial.f90
@@ -1,5 +1,5 @@
 module mod_initial
-    use iso_fortran_env, only: int32, real32
+    use iso_fortran_env, only: int32, real64
     implicit none
     private
     public :: set_gaussian
@@ -7,9 +7,9 @@ module mod_initial
 contains
     pure subroutine set_gaussian(x,  icenter, decay)
         ! Sets x as the graph of a Gaussian with given center and decay
-        real(real32), intent(in out) :: x(:)
+        real(real64), intent(in out) :: x(:)
         integer(int32), intent(in) ::  icenter
-        real(real32), intent(in) :: decay
+        real(real64), intent(in) :: decay
         integer(int32) :: i
         do concurrent(i = 1:size(x))
             x(i) = exp(-decay * (i - icenter)**2)
